# docs/50-api/openapi/sgtm-v1.yaml
...
openapi: 3.1.0
info:
  title: SGTM — Sistema de Gestion Tributaria Municipal
  version: 1.0.0-borrador
  description: |
    Contrato de la API del SGTM, derivado de las pantallas del prototipo de
    interfaz. Una operacion por opcion del menu.

    El identificador de municipalidad NO viaja en ninguna ruta, parametro ni
    cuerpo: sale del claim `municipalidad_id` del token validado (ADR-0005).
    Una peticion que lo mande por otro camino no lo consigue.

    Los importes son cadenas decimales, nunca numeros de coma flotante
    (RNF-055), y toda cifra de deuda indica a que fecha esta actualizada
    (RNF-075).
servers:
  - url: /api/v1
    description: Camino base
security:
  - tokenDeAcceso: []
...
paths:
...
  "/transito/papeletas":
    get:
      operationId: papeletas
      summary: "Papeletas de infracción de tránsito"
      description: "Papeletas levantadas por el inspector municipal, con el código del Reglamento Nacional de Tránsito, la sanción en porcentaje de UIT y la medida preventiva aplicada."
      tags: ["Tránsito"]
      parameters:
        - name: nroPapeleta
          in: query
          required: false
          description: "Filtro «Nro. Papeleta» de la pantalla"
          schema: { type: string }
        - name: placa
          in: query
          required: false
          description: "Filtro «Placa» de la pantalla"
          schema: { type: string }
        - name: documentoDelInfractor
          in: query
          required: false
          description: "Filtro «Documento del infractor» de la pantalla"
          schema: { type: string }
        - name: desde
          in: query
          required: false
          description: "Filtro «Desde» de la pantalla"
          schema: { type: string }
        - name: hasta
          in: query
          required: false
          description: "Filtro «Hasta» de la pantalla"
          schema: { type: string }
        - name: estado
          in: query
          required: false
          description: "Filtro «Estado» de la pantalla"
          schema: { type: string }
        - name: pagina
          in: query
          required: false
          description: "Pagina que se pide, contada desde 0"
          schema: { type: string }
          example: "0"
        - name: tamano
          in: query
          required: false
          description: "Filas por pagina"
          schema: { type: string }
          example: "20"
        - name: ordenarPor
          in: query
          required: false
          description: "Campo por el que se ordena, en camelCase"
          schema: { type: string }
        - name: direccion
          in: query
          required: false
          description: "ASCENDENTE | DESCENDENTE"
          schema: { type: string }
          example: "ASCENDENTE"
      responses:
        200:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
...
  "/transito/constancias-libres":
    post:
      operationId: transito_constancia_libre
      summary: "Constancia libre de infracciones"
      description: >-
        Emite el documento con el que la municipalidad acredita que un vehículo no registra
        papeletas de tránsito pendientes (#53, RF-068). `verificadaAl` es el día al que se
        acredita —no el de emisión— y viaja impreso en el papel (regla 9, RNF-075). Si a esa fecha
        hay una sola papeleta pendiente responde **409** con los números que lo impiden. La
        respuesta es el archivo, en el formato pedido.
      tags: ["Tránsito"]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        201:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
...
  "/consultas/constancias/no-adeudo":
    get:
      operationId: constancia
      summary: "Constancia de no adeudo"
      description: "Vista previa del documento que se entrega al contribuyente. Se imprime con el mismo formato en papel membretado."
      tags: ["Consultas"]
      parameters:
        - name: codContribuyente
          in: query
          required: false
          description: "Filtro «Cod. Contribuyente» de la pantalla"
          schema: { type: string }
        - name: fecha
          in: query
          required: false
          description: "Fecha de corte; sin ella, la de hoy"
          schema: { type: string }
      responses:
        200:
          description: Operacion realizada
          content:
            application/json:
              schema: { type: object }
        "403":
          $ref: "#/components/responses/SinMunicipalidad"
        "422":
          $ref: "#/components/responses/ErrorDeValidacion"
...
components:
  securitySchemes:
    tokenDeAcceso:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token OIDC validado. Debe traer el claim `municipalidad_id`; sin el, la
        peticion recibe 403 y no llega al controlador (ADR-0005, RNF-032).
  schemas:
...
    Error:
      type: object
      required: [codigo, mensaje]
      properties:
        codigo: { type: string, example: "DEUDA_YA_CANCELADA" }
        mensaje: { type: string, description: "En castellano; se muestra al usuario" }
        detalles: { type: array, items: { type: string } }
  responses:
    SinMunicipalidad:
      description: |
        El token no identifica una municipalidad. No hay valor por omision ni
        modo sin municipalidad.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ErrorDeValidacion:
      description: La peticion no cumple una regla de negocio
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }